+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * **Breaking change**: Changed positional arguments of
   [HexaPDF::Layout::TextBox::new] and [HexaPDF::Layout::ImageBox::new] to
   keyword arguments for a consistent box initialization interface
+* [HexaPDF::Layout::Box#split] to provide a default implementation that is
+  useful for most subclasses
 * Layout box implementations to provide a `#supports_position_flow?` method that
   indicates whether the box supports flowing its content around other content.
 * `hexapdf info --check` to only check the current version of each object